refactor(web): move collab bootstrap out of setup.web
